+++ custom_components/vaillant_ebus/const.py
@@ -21,8 +21,15 @@ DISCOVERY_PORT = 8888
 DISCOVERY_TIMEOUT = 3
 DISCOVERY_CANDIDATES = ["core-ebusd", "localhost", "127.0.0.1", "homeassistant.local"]
 SENSITIVE_FIELDS: set[str] = {
-    "serial", "keycode", "installer", "bc", "code",
-    "password", "secret", "token", "pin",
+    "serial",
+    "keycode",
+    "installer",
+    "bc",
+    "code",
+    "password",
+    "secret",
+    "token",
+    "pin",
 }
 
 PLATFORMS = [